feat: #1325 Phase 5b \u2014 retire bg-task management trio
The `ListBackgroundTasks` / `CancelTask` / `WaitTask` LLM tools (added
in #996 Layer 2) are superseded by `WaitForMail` + the mailbox bridge
landed in #1336 (`notify_parent_mailbox`): every bg-agent completion
now sends mail to the parent's mailbox, so the model needs exactly one
tool to block on background work \u2014 `WaitForMail`. Listing and explicit
per-task cancellation never had codex/Claude/Gemini equivalents and
were a koda-specific surface from the pre-mailbox era.

This matches the project's rip-and-replace convention for tool-shape
breaking changes (cf. #1163 removing the `background:bool` flag from
`InvokeAgent` outright, no deprecation period).

## What's deleted

- `koda-core/src/tools/bg_task_tools.rs` (1397 lines: 3 tool defs +
  3 tool stub structs + execute() dispatch + JSON snapshot helpers +
  timeout clamping + ~25 tests).
- The `if matches!(tool_name, "ListBackgroundTasks" | "CancelTask" |
  "WaitTask")` early-dispatch branch in `tool_dispatch::execute_one_tool`
  (Layer-2-of-#996 special path; the trio needed
  `Arc<ChildAgentRegistry>` and the caller's spawner identity, neither
  held by `ToolRegistry`). Branch is gone; `bg_agents` and `cancel`
  drop from the local destructure since no other code in the function
  body uses them.
- Catalog registrations: `bg_task_tools::definitions()` loop +
  `boxed(...)` for the 3 stub structs + the 3 names from CANONICAL.
- `META_TOOLS` entries in `skill_scope.rs` (the trio was
  always-available so skill-scoped agents could still cancel/wait
  on bg work \u2014 obsolete now that `WaitForMail` is meta).
- snake_case alias map entries in `tool_normalize.rs`
  (`list_background_tasks`, `list_bg_tasks`, `cancel_task`,
  `wait_task`, `wait_for_task`).

## What's kept

- `TaskId` enum + `parse_task_id()` helper, moved into a new tiny
  module `koda-core/src/tools/task_id.rs` (~95 lines, just the parser
  + 3 tests). Still consumed by `koda-cli/src/tui_bg_tasks.rs` and
  `tui_commands.rs` for the `/cancel <id>` slash command \u2014 the TUI
  needs to parse user-typed IDs and route to the right registry.
- `ChildAgentRegistry`, `BgRegistry`, `ChildTaskSnapshot` etc. \u2014
  these live one layer down, are still populated on every bg-agent
  spawn, drive the TUI `child_activity_overlay`, and feed the
  `notify_parent_mailbox` walk on bg-agent exit. Phase 5b is purely
  a tool-surface retirement; the runtime registries stay.
- TUI rendering for historical `WaitTask` / `ListBackgroundTasks`
  rows: `wait_task_format.rs` (534 lines), the `if name == "WaitTask"`
  branches in `tui_render.rs` and `history_render.rs`. Pre-5b session
  DBs persisted on disk still contain rows with these tool names \u2014
  resuming an old session must render them correctly. Annotated the
  module-doc on `wmat.rs` so the back-compat-only purpose
  is obvious to a future reader.

## Test migrations

- `e2e_agent_test::test_sub_agent_cache_hit_skips_llm`: barrier between
  two `InvokeAgent` calls switches from
  `WaitTask({"task_ids":["agent:1"]})` to `WaitForMail({})`. The
  semantics are identical for this test \u2014 the bg-completion mail from
  #1336's bridge fires AFTER `sub_agent_cache.put`, so `WaitForMail`
  gives the same happens-before edge that pinned the cache hit before.
- `new_tools_test::BUILTIN_TOOLS`: 25 \u2192 22 (drop the 3 retired names).
- `tool_wiring_test::HIGHER_LAYER_DISPATCH`: now empty (the bypass
  list existed exclusively for the bg-task trio; `InvokeAgent` /
  `SpawnAgent` go through `registry.execute()` and return non-Unknown
  sentinel strings, so they don't need exclusion).
- `skill_scope::skill_scope_does_not_block_meta_tools`: pin
  `SpawnAgent` / `SendMessage` / `WaitForMail` instead of the retired
  trio.

Verified:
- cargo test -p koda-core --lib  \u2192  1387 passed (was 1412; \u221225 from
  the deleted bg_task_tools.rs test mod).
- cargo test -p koda-cli --lib  \u2192  634 passed.
- cargo test -p koda-core --tests  \u2192  all integration suites pass.
- cargo clippy --workspace --lib --tests --all-features -- -D warnings
  \u2192  clean.

Follows up: Phase 5b commit 2 will rewrite `InvokeAgent`'s description
to point at `SpawnAgent` and `WaitForMail` (currently still mentions
`WaitTask` repeatedly). DESIGN.md + CHANGELOG come last.
