+++ koda-core/tests/e2e_agent_test.rs
@@ -269,13 +269,17 @@ async fn test_sub_agent_cache_hit_skips_llm() {
         // here, iter 2 (the would-be cache hit) races iter 1's
         // bg agent and almost always wins on Linux — result: cache
         // miss, second spawn, no `cache hit` Info event, test fails.
-        // WaitTask blocks until the bg agent reaches terminal status,
-        // which can only happen *after* `cache.put` runs (same
-        // function body, return Ok comes after the put). Modeling the
-        // serialization explicitly is also more honest — in the new
-        // world, a model that wants dedupe must serialize via
-        // WaitTask, exactly as written here.
-        MockResponse::tool_call("WaitTask", serde_json::json!({"task_ids": ["agent:1"]})),
+        //
+        // Pre-#1325 Phase 5b this test used `WaitTask(["agent:1"])`
+        // for the barrier. Phase 5b retired `WaitTask`; the equivalent
+        // serialization now goes through `WaitForMail` — the mailbox
+        // bridge from #1336 (`notify_parent_mailbox`) sends a
+        // completion mail to the parent the moment the bg agent
+        // exits, AFTER `cache.put` has run (same function body,
+        // return Ok comes after the put). `WaitForMail` blocks the
+        // current turn until that mail arrives, giving us the same
+        // happens-before edge.
+        MockResponse::tool_call("WaitForMail", serde_json::json!({})),
         MockResponse::tool_call(
             "InvokeAgent",
             serde_json::json!({"agent_name": "echo-agent", "prompt": "do the thing"}),
